# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is prototype of aquafarm management service",
        "title": "Aquafarm Management Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v1/farms/{farmID}/ponds/{code}": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update the pond if pond is exist or insert if pond is not exist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ponds"
                ],
                "summary": "Update or insert pond",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Farm ID",
                        "name": "farmID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pond Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Pond Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpsertPondRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PondResponse"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/v1/ponds/{code}": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update the pond if pond is exist or insert if pond is not exist",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ponds"
                ],
                "summary": "Update or insert pond",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Farm ID",
                        "name": "farmID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pond Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Pond Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpsertPondRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PondResponse"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/ponds/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Get pond data by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ponds"
                ],
                "summary": "Get pond data by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pond ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PondResponse"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Delete pond by id and ponds connected to pond id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ponds"
                ],
                "summary": "Delete pond by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pond ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/generateResp.Error"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {},
                                        "message": {
                                            "type": "string"
                                        },
                                        "ok": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.FarmResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "ponds": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PondResponse"
                    }
                },
                "postalCode": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "village": {
                    "type": "string"
                }
            }
        },
...
        "dto.PondResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "depth": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "farm": {
                    "$ref": "#/definitions/dto.FarmResponse"
                },
                "farmId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "long": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "updatedBy": {
                    "type": "string"
                },
                "wide": {
                    "type": "number"
                }
            }
        },
...
        "dto.UpsertPondRequest": {
            "type": "object",
            "required": [
                "code",
                "depth",
                "description",
                "loginData",
                "long",
                "wide"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "depth": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "farmCode": {
                    "type": "string"
                },
                "farmId": {
                    "type": "integer"
                },
                "loginData": {},
                "long": {
                    "type": "number"
                },
                "wide": {
                    "type": "number"
                }
            }
        },
        "generateResp.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "generateResp.Success": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}